remove translator
