+++ lib/surface/live_view.ex
@@ -32,9 +32,7 @@ defmodule Surface.LiveView do
 
   defmacro __using__(opts) do
     quote do
-      use Surface.BaseComponent,
-        translator: Surface.Translator.LiveViewTranslator,
-        type: unquote(__MODULE__)
+      use Surface.BaseComponent, type: unquote(__MODULE__)
 
       use Surface.API, include: [:property, :data]
       import Phoenix.HTML
