--- lib/surface/translator/component_translator.ex
@@ -1,12 +1,0 @@
-defmodule Surface.Translator.ComponentTranslator do
-  @moduledoc false
-
-  alias Surface.Translator
-
-  @behaviour Translator
-
-  @impl true
-  def translate(node, caller) do
-    Surface.Translator.LiveComponentTranslator.translate(node, caller)
-  end
-end
