--- lib/surface/translator/slotable_translator.ex
@@ -1,13 +1,0 @@
-defmodule Surface.Translator.SlotableTranslator do
-  @moduledoc false
-
-  alias Surface.Translator
-  alias Surface.Translator.ComponentTranslator
-
-  @behaviour Translator
-
-  @impl true
-  def translate(node, caller) do
-    ComponentTranslator.translate(node, caller)
-  end
-end
